+++ custom_patterns.py
@@ -1,7 +1,15 @@
 # custom_patterns.py
-# This file stores user-defined regex patterns.
+# Version: 25.1.1
+# Last modified: 2025-07-02
+# This file stores user-defined regex patterns for model number and QA number detection.
 
 MODEL_PATTERNS = [
+    r'\bDP\b',
+    r'\bM\d+idn\b',
+    r'\bM\d+idnf\b',
+    r'\bDevice Manager\b',
+    r'\bVi\d+\b',
+    r'\bKM-\d+\b',
     r'\bP\d+dn\b',
     r'\bFS\-\d+DN\b',
 ]